+++ level.h
@@ -12,18 +12,18 @@ class Level
 {
 public:
 	Level();
-	Level(size_t h, size_t w, size_t num_balls);
+	Level(uint32_t h, uint32_t w, size_t num_balls);
 
 	void update_level();
 
 	Collision bCollision(Position p);
 
-	Cell get_level_pixel(size_t h, size_t w);
+	Cell get_level_pixel(uint32_t h, uint32_t w);
 
 	std::vector<Ball> get_balls();
 private:
-	size_t height;
-	size_t width;
+	uint32_t height;
+	uint32_t width;
 	std::vector<Ball> balls;
 	std::vector<Cell> level;
 };
